% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data()
}
\value{
A list containing the objects that {stargazer} can deal with.
}
\description{
A function that captures the objects that work with {stargazer}, such as `lm` objects. It is necessary since the user has to choose the objects he/she wants to put in the table. Function coming from [this answer](https://community.rstudio.com/t/how-can-i-access-the-users-global-environment/62688/2)
}
\examples{
# make a random regression and create a random dataframe
regression <- lm(mpg ~ drat + hp + disp, data = datasets::mtcars)
test <- mtcars
# store only the objects whose class is "lm" in a list
get_data()
}
