% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_data.R
\docType{data}
\name{models_supported_stargazer}
\alias{models_supported_stargazer}
\title{List of models supported by {stargazer}}
\format{
An object of class \code{character} of length 69.
}
\usage{
models_supported_stargazer
}
\value{
All the models supported by {stargazer}, see the list here: https://rdrr.io/cran/stargazer/man/stargazer_models.html
}
\description{
List of models supported by {stargazer}
}
\examples{
NULL
}
\keyword{datasets}
