% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Title}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{something}

\item{output}{something else}

\item{session}{something else}
}
\value{

}
\description{
Title
}
\examples{
NULL
}
